♻️ Refactor submission APIs to repository
