+++ src/params/indomitableBoss.ts
@@ -1,10 +1,10 @@
 import type { ParamMatcher } from '@sveltejs/kit';
 
-const indomitableBosses: { [key: string]: string } = {
-	nexaelio: 'Nex Aelio',
-	renusretem: 'Renus Retem',
-	amskvaris: 'Ams Kvaris',
-	nilsstia: 'Nils Stia'
+const indomitableBosses: { [key: string]: boolean } = {
+	nexaelio: true,
+	renusretem: true,
+	amskvaris: true,
+	nilsstia: true
 };
 
 export const match = ((param) => {
